Moved offset_t to separate module.
