+++ CMakeLists.txt
@@ -84,6 +84,7 @@ target_sources(PinkPearl PRIVATE
 	src/util/bit.c
 	src/util/extent.c
 	src/util/file_io.c
+	src/util/offset.c
 	src/util/string_array.c
 	src/util/time.c
 )
