+++ src/util/extent.h
@@ -10,13 +10,4 @@ typedef struct extent_t {
 
 uint64_t extent_area(const extent_t extent);
 
-typedef struct offset_t {
-	int32_t x;
-	int32_t y;
-} offset_t;
-
-offset_t offset_add(const offset_t a, const offset_t b);
-
-offset_t offset_subtract(const offset_t a, const offset_t b);
-
 #endif	// EXTENT_H
